backend_vercel/requirements.txt
fastapi==0.115.6
SQLAlchemy==2.0.36
pandas==2.2.3
httpx==0.28.1
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.20
